# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ptydeck API
  version: 0.1.0
  description: OpenAPI contract for ptydeck session lifecycle and terminal control.
servers:
  - url: /api/v1
paths:
...
  /shares:
    get:
      operationId: listShares
      summary: List persisted read-only share links.
      responses:
        '200':
          description: Share link list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShareLink'
        '426':
          $ref: '#/components/responses/TlsRequired'
    post:
      operationId: createShareLink
      summary: Create a read-only spectator share link for a session or deck.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareLinkRequest'
      responses:
        '201':
          description: Share link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLink'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Target not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
  /shares/{shareId}:
    parameters:
      - $ref: '#/components/parameters/shareId'
    get:
      operationId: getShareLink
      summary: Get share link metadata by id.
      responses:
        '200':
          description: Share link found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLink'
        '404':
          description: Share link not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
  /shares/{shareId}/revoke:
    parameters:
      - $ref: '#/components/parameters/shareId'
    post:
      operationId: revokeShareLink
      summary: Revoke a read-only spectator share link.
      responses:
        '200':
          description: Share link revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLink'
        '404':
          description: Share link not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
  /custom-commands:
    get:
      operationId: listCustomCommands
      summary: List persisted custom commands.
      parameters:
        - $ref: '#/components/parameters/customCommandScope'
        - $ref: '#/components/parameters/customCommandSessionId'
      responses:
        '200':
          description: Custom command list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomCommand'
        '426':
          $ref: '#/components/responses/TlsRequired'
  /custom-commands/{commandName}:
    parameters:
      - $ref: '#/components/parameters/commandName'
    get:
      operationId: getCustomCommand
      summary: Get custom command by name.
      parameters:
        - $ref: '#/components/parameters/customCommandScope'
        - $ref: '#/components/parameters/customCommandSessionId'
      responses:
        '200':
          description: Custom command found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomCommand'
        '404':
          description: Custom command not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Custom command name resolves to multiple scoped commands
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
    put:
      operationId: upsertCustomCommand
      summary: Create or overwrite custom command by name.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCustomCommandRequest'
      responses:
        '200':
          description: Custom command upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomCommand'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Custom command name conflict or limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
    delete:
      operationId: deleteCustomCommand
      summary: Delete custom command by name.
      parameters:
        - $ref: '#/components/parameters/customCommandScope'
        - $ref: '#/components/parameters/customCommandSessionId'
      responses:
        '204':
          description: Custom command deleted
        '404':
          description: Custom command not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Custom command name resolves to multiple scoped commands
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
  /decks:
    get:
      operationId: listDecks
      summary: List decks.
      responses:
        '200':
          description: Deck list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deck'
        '426':
          $ref: '#/components/responses/TlsRequired'
    post:
      operationId: createDeck
      summary: Create deck.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeckRequest'
      responses:
        '201':
          description: Deck created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deck'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Deck id already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
  /decks/{deckId}:
    parameters:
      - $ref: '#/components/parameters/deckId'
    get:
      operationId: getDeck
      summary: Get deck by id.
      responses:
        '200':
          description: Deck found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deck'
        '404':
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
    patch:
      operationId: updateDeck
      summary: Update deck metadata/settings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeckRequest'
      responses:
        '200':
          description: Deck updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deck'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
    delete:
      operationId: deleteDeck
      summary: Delete deck by id.
      parameters:
        - name: force
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Deck deleted
        '400':
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Deck delete conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
  /decks/{deckId}/sessions/{sessionId}:move:
    parameters:
      - $ref: '#/components/parameters/deckId'
      - $ref: '#/components/parameters/sessionId'
    post:
      operationId: moveSessionToDeck
      summary: Move session to deck.
      responses:
        '204':
          description: Move accepted
        '404':
          description: Deck or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
  /layout-profiles:
    get:
      operationId: listLayoutProfiles
      summary: List persisted layout profiles.
      responses:
        '200':
          description: Layout profile list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LayoutProfile'
        '426':
          $ref: '#/components/responses/TlsRequired'
    post:
      operationId: createLayoutProfile
      summary: Create a persisted layout profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLayoutProfileRequest'
      responses:
        '201':
          description: Layout profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LayoutProfile'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Layout profile id already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
  /layout-profiles/{profileId}:
    parameters:
      - $ref: '#/components/parameters/profileId'
    get:
      operationId: getLayoutProfile
      summary: Get layout profile by id.
      responses:
        '200':
          description: Layout profile found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LayoutProfile'
        '404':
          description: Layout profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
    patch:
      operationId: updateLayoutProfile
      summary: Update persisted layout profile metadata/layout.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLayoutProfileRequest'
      responses:
        '200':
          description: Layout profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LayoutProfile'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Layout profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
    delete:
      operationId: deleteLayoutProfile
      summary: Delete persisted layout profile by id.
      responses:
        '204':
          description: Layout profile deleted
        '404':
          description: Layout profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
  /connection-profiles:
    get:
      operationId: listConnectionProfiles
      summary: List persisted connection profiles.
      responses:
        '200':
          description: Connection profile list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectionProfile'
        '426':
          $ref: '#/components/responses/TlsRequired'
    post:
      operationId: createConnectionProfile
      summary: Create a persisted connection profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnectionProfileRequest'
      responses:
        '201':
          description: Connection profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionProfile'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Connection profile id already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
  /connection-profiles/{profileId}:
    parameters:
      - $ref: '#/components/parameters/profileId'
    get:
      operationId: getConnectionProfile
      summary: Get connection profile by id.
      responses:
        '200':
          description: Connection profile found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionProfile'
        '404':
          description: Connection profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
    patch:
      operationId: updateConnectionProfile
      summary: Update persisted connection profile metadata/launch settings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConnectionProfileRequest'
      responses:
        '200':
          description: Connection profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionProfile'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Connection profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
    delete:
      operationId: deleteConnectionProfile
      summary: Delete persisted connection profile by id.
      responses:
        '204':
          description: Connection profile deleted
        '404':
          description: Connection profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
...
  /ssh-trust-entries/{entryId}:
    parameters:
      - name: entryId
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteSshTrustEntry
      summary: Delete a persisted SSH host-key trust entry.
      responses:
        '204':
          description: Trust entry deleted
        '404':
          description: Trust entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '426':
          $ref: '#/components/responses/TlsRequired'
...
components:
  parameters:
...
    commandName:
      name: commandName
      in: path
      required: true
      schema:
        type: string
    customCommandScope:
      name: scope
      in: query
      required: false
      schema:
        type: string
        enum:
          - global
          - project
          - session
    customCommandSessionId:
      name: sessionId
      in: query
      required: false
      schema:
        type: string
...
    shareId:
      name: shareId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    CreateDeckRequest:
      type: object
      required:
        - name
      properties:
        id:
          type: string
        name:
          type: string
        settings:
          $ref: '#/components/schemas/DeckSettings'
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
...
    CreateShareLinkRequest:
      type: object
      required:
        - targetType
        - targetId
      properties:
        targetType:
          type: string
          enum:
            - session
            - deck
        targetId:
          type: string
        expiresInSeconds:
          type: integer
          format: int64
    ShareLink:
      type: object
      required:
        - id
        - targetType
        - targetId
        - permissionMode
        - createdAt
        - updatedAt
        - expiresAt
        - revokedAt
        - creatorSubject
        - creatorTenantId
        - active
      properties:
        id:
          type: string
        targetType:
          type: string
          enum:
            - session
            - deck
        targetId:
          type: string
        permissionMode:
          type: string
          enum:
            - read_only
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        expiresAt:
          type: integer
          format: int64
        revokedAt:
          oneOf:
            - type: integer
              format: int64
            - type: 'null'
        creatorSubject:
          type: string
        creatorTenantId:
          type: string
        active:
          type: boolean
        joinUrl:
          type: string
  responses:
    TlsRequired:
      description: TLS is required for this endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'